prevent fetching data if no session and fix member form
